==> Analysis 2 (BECs vs SCECs)/04_merge_BA4andCSC.Rmd <==
---
title: "Merge BA4 and CSC pure ECs"
author: "B180131 dissertation"
date: "30/05/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

This is the scRNA-seq analysis workflow for comparing human brain endothelial cells vs human spinal cord endothelial cells. This allows for better understanding of how neurovascular functions are contributed by selected populations of EC populations. Required packages involve Seurat and Bioconductor. 

* This script attempts to merge "pure" endothelial cell populations from BA4, CB, and CSC, saves this to an RDS object.

```{r, include=FALSE}
load_lib <- "Load Libraries"
```

```{r load_lib, message=FALSE}
library(Seurat) 
library(dplyr)
library(scater)
library(SingleCellExperiment)
library(here)
library(ggplot2)
library(scran)
library(stringr)
library(clustree)
library(cowplot)
```


```{r, include=FALSE}
load_data <- "Load BA4 & CSC pure EC populations"
```
## `r load_data`
* Load the QC-ed and normalised pure EC populations of BA4 and CSC from Luise's dataset.


```{r load_data, echo=TRUE}
BA4 <- readRDS(here("data/BA4_SubAnnot.RDS")) #Load entire dataset
BA4$sample <- rep("Seeker_BA4", ncol(BA4))
DimPlot(BA4, group.by = "subtype")
BA4_sce <- as.SingleCellExperiment(BA4) #Convert to SCE object

CB <- readRDS(here("data/CB_SubAnnot.RDS")) #Load entire dataset
CB$sample <- rep("Seeker_CB", ncol(CB))
DimPlot(CB, group.by = "subtype")
CB_sce <- as.SingleCellExperiment(CB) #Convert to SCE object

CSC <- readRDS(here("data/CSC_SubAnnot.RDS")) #Load entire dataset
CSC$sample <- rep("Seeker_CSC", ncol(CSC))
DimPlot(CSC, group.by = "subtype")
CSC_sce <- as.SingleCellExperiment(CSC) #Convert to SCE object
```


```{r}
plot_BA4 <- FeaturePlot(BA4, features = c("CLDN5", "FLT1"), pt.size = 0.7) 

plot_BA4 <- plot_grid(ggdraw() + draw_label("BA4 (Motor cortex) endothelial cells", size = 15, fontface = "bold"), plot_BA4, ncol = 1, rel_heights = c(0.1,1))

plot_CB <- FeaturePlot(CB, features = c("CLDN5", "FLT1"), pt.size = 0.7) 

plot_CB <- plot_grid(ggdraw() + draw_label("CB (Cerebellum) endothelial cells", size = 15, fontface = "bold"), plot_CB, ncol = 1, rel_heights = c(0.1,1))

plot_CSC <- FeaturePlot(CSC, features = c("CLDN5", "FLT1"), pt.size = 0.7) 

plot_CSC <- plot_grid(ggdraw() + draw_label("CSC (Cervical spinal cord) endothelial cells", size = 15, fontface = "bold"), plot_CSC, ncol = 1, rel_heights = c(0.1,1))

gridExtra::grid.arrange(plot_BA4, plot_CB, plot_CSC, nrow = 3)
```


```{r, include=FALSE}
merge_data <- "Merge BA4 & CSC EC datasets"
```
## `r merge_data`
* Here, we use the merge function and "add ids" of the two SeuratObjects to prevent barcode overlap


```{r}
BEC <- merge(BA4, CB, add.cell.ids = c("BA4", "CB"))
BEC_sce <- as.SingleCellExperiment(BEC)
comb_EC <- merge(BA4, c(CB,CSC), add.cell.ids = c("BA4", "CB", "CSC"))
comb_EC_sce <- as.SingleCellExperiment(comb_EC) #Convert SCE object
```


```{r, include=FALSE}
QC_data <- "QC Visualisation of BA EC populations"
```
## `QC_data`
It is worthy to first visualize the QC of each EC population.

* Plot highest expressed genes to identify EC markers 
* Visualize QC metrics using Scater
* Violin plot to identify MAD outliers
* Scatter plot on metrics ratio


#Plot highest expressed genes in each EC populations
```{r QC_data, message=FALSE, fig.width=10, fig.height= 5}
exp_plot1 <- plotHighestExprs(BA4_sce, exprs_values = "counts")

exp_plot2 <- plotHighestExprs(CB_sce, exprs_values = "counts")

exp_plot3 <- plotHighestExprs(CSC_sce, exprs_values = "counts")

gridExtra::grid.arrange(exp_plot1, exp_plot2, exp_plot3, ncol = 3)
```


#Plot highest expressed genes in each EC populations
```{r, message=FALSE}
EC_markers_1 <- plotExpression(BEC_sce, exprs_values = "logcounts", features = c("VWF",
                              "FLT1",
                              "CLDN5",
                              "ABCB1",
                              "ST6GALNAC3",
                              "ELOVL7",
                              "THSD4",
                              "PECAM1",
                              "MECOM"
                              )) + ggtitle("BEC")+ ggtitle("SCEC") + ggtitle("BEC") + theme(plot.title = element_text(hjust = 0.5))

EC_markers_2 <- plotExpression(CSC_sce, exprs_values = "logcounts", features = c("VWF",
                              "FLT1",
                              "CLDN5",
                              "ABCB1",
                              "ST6GALNAC3",
                              "ELOVL7",
                              "THSD4",
                              "PECAM1",
                              "MECOM"
                              )) + ggtitle("SCEC") + ggtitle("SCEC") + theme(plot.title = element_text(hjust = 0.5))
EC_markers_1
EC_markers_2
gridExtra::grid.arrange(EC_markers_1, EC_markers_2, nrow = 2)
```


#Identify outliers and visualize QC metrics using Scater
```{r, message = FALSE,  fig.width= 12, fig.height= 6}
out_mito <- isOutlier(comb_EC_sce$percent.mito, log = TRUE, type = "higher", nmads = 3)

out_feature <- isOutlier(comb_EC_sce$nFeature_RNA, log = TRUE, type = "both", nmads = 3)

out_lib <- isOutlier(comb_EC_sce$nCount_RNA, log = TRUE, type = "both", nmads = 3)

#Plot violin
plot_mito <- plotColData(comb_EC_sce, x = "Tissues", y = "percent.mito", colour_by = I(out_mito)) + theme(axis.text.x = element_text(size = 20),
                  axis.text.y = element_text(size = 20),
                  axis.title.x = element_text(size = 20),
                  axis.title.y = element_text(size = 20),
                  legend.text = element_text(size = 15)) +
        guides(colour = guide_legend(override.aes = list(size=4)))

plot_feature <- plotColData(comb_EC_sce, x = "Tissues", y = "nFeature_RNA", colour_by = I(out_feature)) + theme(axis.text.x = element_text(size = 20),
                  axis.text.y = element_text(size = 20),
                  axis.title.x = element_text(size = 20),
                  axis.title.y = element_text(size = 20),
                  legend.text = element_text(size = 15)) +
        guides(colour = guide_legend(override.aes = list(size=4)))


plot_lib <- plotColData(comb_EC_sce, x = "Tissues", y = "nCount_RNA", colour_by = I(out_lib)) + theme(axis.text.x = element_text(size = 20),
                  axis.text.y = element_text(size = 20),
                  axis.title.x = element_text(size = 20),
                  axis.title.y = element_text(size = 20),
                  legend.text = element_text(size = 15)) +
        guides(colour = guide_legend(override.aes = list(size=4)))


gridExtra::grid.arrange(plot_mito, plot_feature, plot_lib, ncol = 3) 
```


#Scatter plot (QC metric ratio)
```{r, message = FALSE, fig.width= 12, fig.height= 6}
lib_feature <- plotColData(comb_EC_sce, x ="nCount_RNA", y = "nFeature_RNA", colour_by = "Tissues")
lib_mito <- plotColData(comb_EC_sce, x ="nCount_RNA", y = "percent.mito", colour_by = "Tissues")
gridExtra::grid.arrange(lib_feature, lib_mito, ncol=2) #Plot scatter plot of QC metrics
```


```{r}
saveRDS(object = comb_EC, file = "/Users/bonocheong/Desktop/Research_Project/Datasets/Dissertation/data/merged_EC.RDS")
```
